+++ includes/plugin/class-core.php
@@ -17,6 +17,7 @@ use Traffic\System\I18n;
 use Traffic\System\Assets;
 use Traffic\Library\Libraries;
 use Traffic\System\Nag;
+use Traffic\System\Role;
 
 /**
  * The core plugin class.
@@ -108,6 +109,8 @@ class Core {
 		$this->loader->add_action( 'admin_notices', $nag, 'display' );
 		$this->loader->add_action( 'wp_ajax_hide_traffic_nag', $nag, 'hide_callback' );
 		$this->loader->add_action( 'wp_ajax_traffic_get_stats', 'Traffic\Plugin\Feature\AnalyticsFactory', 'get_stats_callback' );
+		$this->loader->add_filter( 'myblogs_blog_actions', $plugin_admin, 'blog_action', 10, 2 );
+		$this->loader->add_filter( 'manage_sites_action_links', $plugin_admin, 'site_action', 10, 3 );
 	}
 
 	/**